Added wallet service features and send money in transaction service with integration of jwt security

+++ user-service/src/main/java/com/user/dto/UserDto.java
@@ -7,7 +7,7 @@ import lombok.Data;
 public class UserDto {
 private int userId;
 private String userName;
-private int phoneNo;
+private long phoneNo;
 private String address;
 private String email;
 private String otp;
